02_model/pyproject.toml
[project]
name = "02-model"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.4",
    "huggingface-hub>=0.29.3",
    "langchain>=0.3.20",
    "langchain-anthropic>=0.3.10",
    "langchain-core>=0.3.45",
    "langchain-deepseek>=0.1.2",
    "langchain-google-genai>=2.0.10",
    "langchain-huggingface>=0.1.2",
    "langchain-openai>=0.3.9",
    "numpy>=2.2.4",
    "openai>=1.66.3",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.6.1",
    "transformers>=4.49.0",
]
